pyproject.toml
[tool.poetry]
name = "bukichi-bot"
version = "0.1.0"
description = "Discord bukichi bot"
authors = ["otajisan <mtaji@morningcode.io>"]
readme = "README.md"
packages = [{include = "bukichi_bot"}]

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["voice"], version = "^2.2.2"}
pillow = "^10.1.0"
regex = "^2023.12.25"
requests = "^2.31.0"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
